# 028. Swagger (API UI)/api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made with Express and documented with Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to login.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignIn"
              }
            }
          }
        }
      }
    },
...
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to signup.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUp"
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to reset password.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reset"
              }
            }
          }
        }
      }
    },
...
    "/update-password": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to update password.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassword"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SignUp": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "anv"
          },
          "email": {
            "type": "string",
            "example": "anv@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "re_password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "re_password"
        ],
        "xml": {
          "name": "SignUp"
        }
      },
      "SignIn": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anb@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "SignIn"
        }
      },
      "Reset": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anv@gmai.com"
          }
        },
        "required": [
          "email"
        ],
        "xml": {
          "name": "Reset"
        }
      },
...
    }
  }
}